refactor: rename weather.py to provider.py
