+++ merino/providers/weather/provider.py
